+++ node/src/main/kotlin/net/corda/node/services/transactions/ValidatingNotaryFlow.kt
@@ -1,4 +1,4 @@
-package net.corda.flows
+package net.corda.node.services.transactions
 
 import co.paralleluniverse.fibers.Suspendable
 import net.corda.core.contracts.TransactionVerificationException
@@ -8,6 +8,7 @@ import net.corda.core.node.services.UniquenessProvider
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.WireTransaction
 import net.corda.core.utilities.unwrap
+import net.corda.flows.*
 import java.security.SignatureException
 
 /**
